separated shared data layer
