+++ bankapp/build.gradle
@@ -1,3 +1,3 @@
 dependencies {
-    implementation(project(":webapp"))
+    implementation(project(":data"))
 }
