+++ bankapp/src/main/java/am/cs322/bank/Cs322HomeworkApplicationBank.java
@@ -2,7 +2,10 @@ package am.cs322.bank;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-@SpringBootApplication(scanBasePackages = {"am.cs322.bank"})
+import org.springframework.context.annotation.ComponentScan;
+
+@SpringBootApplication
+@ComponentScan( {"am.cs322.bank", "am.cs322.data"})
 public class Cs322HomeworkApplicationBank {
 
     public static void main(String[] args) {
