+++ bankapp/src/main/java/am/cs322/bank/business/api/BankService.java
@@ -1,6 +1,6 @@
 package am.cs322.bank.business.api;
 
-import am.cs322.bank.domain.model.BankAccount;
+import am.cs322.data.model.BankAccount;
 import java.math.BigDecimal;
 
 public interface BankService {
