+++ data/src/main/java/am/cs322/data/BankAccountRepository.java
@@ -1,8 +1,9 @@
-package am.cs322.bank.data;
+package am.cs322.data;
 
-import am.cs322.bank.domain.model.BankAccount;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import am.cs322.data.model.BankAccount;
 
 @Repository
 public interface BankAccountRepository extends JpaRepository<BankAccount, Long> {
