+++ webapp/src/main/java/am/cs322/user/presentation/api/UserController.java
@@ -1,7 +1,8 @@
 package am.cs322.user.presentation.api;
 
 
-import am.cs322.user.domain.model.UserDTO;
+
+import am.cs322.data.model.UserDTO;
 import am.cs322.user.presentation.api.model.CreateUserRequest;
 
 public interface UserController {
