+++ webapp/src/main/java/am/cs322/user/presentation/impl/UserControllerImpl.java
@@ -2,13 +2,13 @@ package am.cs322.user.presentation.impl;
 
 import am.cs322.user.business.api.UserService;
 
-import am.cs322.user.domain.model.UserDTO;
 import am.cs322.user.presentation.api.UserController;
 import am.cs322.user.presentation.api.model.CreateUserRequest;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import am.cs322.data.model.UserDTO;
 
 @RestController
 @RequestMapping("/v1/users")
